content/canvas/test/test_toDataURL.zerosize.html

<!DOCTYPE HTML>
<title>Canvas test: toDataURL.zerosize</title>
<!-- Testing: toDataURL on zero-size canvas returns 'data:,' -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="0" height="0"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var data = canvas.toDataURL();
todo(data === 'data:,', "data === 'data:,'");

SimpleTest.finish();

});
</script>
